Initial _code_ commit...
Moving things over and about to start the "fork".  We know that
this stuff cowpiles instead of compiles, partly because of a
drift in a kernel source header file (Seriously, Pinchart?)in a
userland compile that will need some cleanup.  We're going to,
remove the requirements for Kernel source header files as one of
the first things...right after the patches go in.

(That's next past making a CMake project file so we can build this
the "right" way by hand or by cross-compile in at least OE...)