Cleanups for build.
This compiles clean from start to finish.  We moved
a stripped version of uvc.h from the _kernel_sources_ to
this project because, to be honest, this was a no-no
anyhow.  There's a bit of a reason for the move- this
really, really needs to be a proper test-tool/server
for this gadget feature.