Provisionally adding H.264 encoding support.
Ideally, the rules should be similar enough to MJPEG to be able
to just simply extend most of the support there and turn on the
mode.  This may not work as expected since I'm just blindly
guessing right now- so this grab the prior commit for now.
(This is getting checked in so I can have a line in the sand
that can be kept, morphed against, etc.)